# openapi.yaml
openapi: 3.0.0
info:
    title: TraderJoe API
    description: Easily retrieve info on the state of Trader Joe.
    version: 0.0.1
basePath: /v1
servers:
    - url: https://api.traderjoexyz.com/v1
      description: Production
paths:
...
    /{poolId}/liquidity:
        get:
            summary: Liquidity/TVL of TraderJoe farm pool
            tags:
                - pool
            parameters:
                - name: poolId
                  in: path
                  required: true
                  description: ID of pool
                  example: '3'
                  schema:
                      type: string
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: string
                                        example: '30009819328012'
                                    status:
                                        $ref: '#/components/schemas/status'
    /{poolId}/apr:
        get:
            summary: APR of TraderJoe farm pool
            tags:
                - pool
            parameters:
                - name: poolId
                  in: path
                  required: true
                  description: ID of pool
                  example: '3'
                  schema:
                      type: string
                - name: period
                  description: Sample period to get APR
                  in: query
                  required: false
                  example: '1mo'
                  schema:
                      type: string
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: string
                                        example: '30009819328012'
                                    status:
                                        $ref: '#/components/schemas/status'
    /{poolId}/apr/bonus:
        get:
            summary: Bonus APR of TraderJoe farm pool
            tags:
                - pool
            parameters:
                - name: poolId
                  in: path
                  required: true
                  description: ID of pool
                  example: '3'
                  schema:
                      type: string
                - name: period
                  description: Sample period to get APR
                  in: query
                  required: false
                  example: '1mo'
                  schema:
                      type: string
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: string
                                        example: '30009819328012'
                                    status:
                                        $ref: '#/components/schemas/status'
    /{pairAddress}/liquidity:
        get:
            summary: TVL/Liquidity of TraderJoe pair
            tags:
                - pair
            parameters:
                - name: pairAddress
                  in: path
                  required: true
                  description: Address of pair
                  example: '0x113f413371fc4cc4c9d6416cf1de9dfd7bf747df'
                  schema:
                      type: string
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: string
                                        example: '30009819328012'
                                    status:
                                        $ref: '#/components/schemas/status'
...
    /{pairAddress}/apr:
        get:
            summary: APR of TraderJoe pair
            tags:
                - pair
            parameters:
                - name: pairAddress
                  in: path
                  required: true
                  description: Address of pair
                  example: '0x113f413371fc4cc4c9d6416cf1de9dfd7bf747df'
                  schema:
                      type: string
                - name: period
                  description: Sample period to get APR
                  in: query
                  required: false
                  example: '1mo'
                  schema:
                      type: string
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: string
                                        example: '30009819328012'
                                    status:
                                        $ref: '#/components/schemas/status'
...
components:
    schemas:
...
        status:
            type: object
            properties:
                timestamp:
                    type: string
                    example: '1641341612759'
                error_code:
                    type: integer
                    example: 400
                error_message:
                    type: string
                    example: 'Error getting supply'
                error_trace:
                    type: string
                    example: 'at Object.require.extensions.<computed> [as .ts] (D:\Documents\Development\crypto\traderjoe-api\node_modules\ts-node\src\index.ts:1374:12)'
...
    securitySchemes:
        - api_key:
              type: apiKey
              name: api_key
              in: header
        - admin:
              type: http
              scheme: basic